#32 Refactor MoveTree, MoveNode into Go package
