+++ common/src/main/kotlin/codes/nibby/yi/go/GoGameModel.kt
@@ -1,7 +1,5 @@
 package codes.nibby.yi.go
 
-import codes.nibby.yi.common.MoveNode
-import codes.nibby.yi.common.MoveTree
 import codes.nibby.yi.go.rules.GoGameRulesHandler
 import java.util.*
 import kotlin.collections.HashSet
