+++ common/src/main/kotlin/codes/nibby/yi/go/GoMoveHelper.kt
@@ -1,6 +1,5 @@
 package codes.nibby.yi.go
 
-import codes.nibby.yi.common.MoveNode
 import java.util.*
 import kotlin.collections.HashSet
