+++ common/src/main/kotlin/codes/nibby/yi/go/SubmitMoveException.kt
@@ -1,7 +1,5 @@
 package codes.nibby.yi.go
 
-import codes.nibby.yi.exception.GameRulesException
-
 /**
  * An exception thrown to represent a submitted move is illegal and cannot be played.
  */
